Windows needs utf16 encoding for paths
